bqui: make arrange public and give Builder box variables
