+++ src/bqui/include/bqui/widget/builder.h
@@ -7,6 +7,7 @@
 #include "bqui/buildparams.h"
 #include "bqui/simplesizehint.h"
 #include "bqui/sizehint.h"
+#include "bqui/widget/boxvariables.h"
 
 #include <bq/signal/signal.h>
 
@@ -84,12 +85,14 @@ namespace bqui::widget
         template <typename TSignalSizeHint>
         auto setSizeHint(TSignalSizeHint sizeHint) &&
         {
-            return makeBuilder(
+            auto builder = makeBuilder(
                     std::move(*func_),
                     std::move(sizeHint),
                     std::move(buildParams_),
                     std::move(gravity_)
                     );
+            builder.setBoxVariables(box_);
+            return builder;
         }
 
         SizeHintType getSizeHint() const
@@ -97,9 +100,24 @@ namespace bqui::widget
             return sizeHint_->clone();
         }
 
+        /** @brief The four edge variables that name this widget's box to the
+         * constraint solver. Stable for the builder's lifetime and preserved
+         * across a copy, a size-hint change and type erasure. */
+        BoxVariables const& getBoxVariables() const
+        {
+            return box_;
+        }
+
+        /** @brief Adopts @p box as this widget's solver box, so a transformed
+         * builder keeps the identity of the one it came from. */
+        void setBoxVariables(BoxVariables box)
+        {
+            box_ = std::move(box);
+        }
+
         auto setBuildParams(BuildParams params) &&
         {
-            return makeBuilder([params=std::move(buildParams_),
+            auto builder = makeBuilder([params=std::move(buildParams_),
                     func=std::move(func_)](BuildParams oldParams, auto size)
                 {
                     return (*func)(params, std::move(size)).setParams(oldParams);
@@ -108,6 +126,8 @@ namespace bqui::widget
                 std::move(params),
                 std::move(gravity_)
                 );
+            builder.setBoxVariables(box_);
+            return builder;
         }
 
         BuildParams const& getBuildParams() const
@@ -134,12 +154,14 @@ namespace bqui::widget
 
         operator BuilderBase() &&
         {
-            return BuilderBase(
+            BuilderBase base(
                     std::move(*func_),
                     std::move(*sizeHint_),
                     std::move(buildParams_),
                     std::move(gravity_)
                     );
+            base.setBoxVariables(box_);
+            return base;
         }
 
     protected:
@@ -148,6 +170,7 @@ namespace bqui::widget
         BuildParams buildParams_;
         bq::signal::AnySignal<avg::Vector2f> gravity_ =
             bq::signal::constant(avg::Vector2f(0.5f, 0.5f));
+        BoxVariables box_;
     };
 
     struct AnyBuilder : Builder<std::function<widget::AnyElement(
